bug: Resolved errors when viewing on mobile
This commit addresses the misalignment of certain elements on the page
by readjusting the fluid containers. Some changes were also made to
the expertise descriptions.

Prior to this commit, some elements were laying off the viewport
and causing alignment issues with the whole page. This is no longer
an issue now that I placed them inside fluid containers.