feat: Improved styling of hero section
This commit provides a friendly definition of my name, making the name
look like a dictionary entry. The font sizes for the whole heading
were also adjusted to make them more appealing and scale better when
orientation and device widths are changed.

Prior to this fix, the hero section would look really cluttered, and
certain elements would seem out of place when a user would turn their
mobile devices to landscape orientation. This is no longer an issue.